Refactor tests to multiple files
Previously all tests were hanging out in tests.py and it was topping
nearly 1700 lines and sort of disorganized.

This commit pulls them out with the following general structure:

__init__ : generic utility functions
audit    : tests related to auditing, pending and automatic rep
models   : a catch-all for generic model test
search   : tests for filter parsing and searching
views    : tests for views

As OTM2 grows we may have to revisit these files but for the moment it
seems fine.
